+++ hisubmit/src/Client.SharedModels/Features/Festivals/Queries/GetAllDeadLineEventCategory/GetAllDeadLineEventCategoryResponse.cs
@@ -13,9 +13,9 @@ public class GetAllDeadLineEventCategoryResponse
 
     public GetEventCategoryByIdResponse EventCategory { get; set; } = new();
         
-    public int? StudentFee { get; set; }
-    public int? StandardFee { get; set; }
-    public int? GoldFee { get; set; }
+    public double? StudentFee { get; set; }
+    public double? StandardFee { get; set; }
+    public double? GoldFee { get; set; }
 
     public DateTime DeadLineDate { get; set; }
     public string DeadLineName { get; set; }
