+++ hisubmit/src/Domain/Entities/Festivals/DeadlineEventCategory.cs
@@ -4,9 +4,9 @@ namespace HiSubmit.Domain.Entities.Festivals
 {
     public class DeadlineEventCategory:AuditableEntity<int>
     {
-        public int? GoldFee { get; set; }
-        public int? StudentFee { get; set; }
-        public int? StandardFee { get; set; }
+        public double? GoldFee { get; set; }
+        public double? StudentFee { get; set; }
+        public double? StandardFee { get; set; }
 
         public int EventCategoryId { get; set; }
         public int DeadLineId { get; set; }
